# soya-framework-admin/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "title" : "Workshop Server",
    "contact" : {
      "name" : "wenqun.soya@gmail.com"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/bod/create/{businessObject}" : {
      "post" : {
        "tags" : [ "bod" ],
        "operationId" : "create",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/bod/arrays/{businessObject}" : {
      "post" : {
        "tags" : [ "bod" ],
        "operationId" : "arrays",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mappingFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "construction",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/esql-validation/{businessObject}" : {
      "post" : {
        "tags" : [ "bod" ],
        "operationId" : "esqlValidation",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mappingFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/sample-xml/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "sampleXml",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/json-schema/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "jsonSchema",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "jsonSchemaFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/json-sample/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "jsonSample",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "jsonSchemaFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/sample-avro/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "sampleAvro",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/array-annotate/{businessObject}" : {
      "post" : {
        "tags" : [ "bod" ],
        "operationId" : "arrayAnnotate",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mappingFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/json-types/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "jsonTypes",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/bod/avsc/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "avsc",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/xpath-construct/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "xpathConstruct",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mappingFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "construction",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/bod/esql-gen/{businessObject}" : {
      "post" : {
        "tags" : [ "bod" ],
        "operationId" : "esqlGen",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mappingFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "construction",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/bod/construct-annotate/{businessObject}" : {
      "post" : {
        "tags" : [ "bod" ],
        "operationId" : "constructAnnotate",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "mappingFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/bod/schema/{businessObject}" : {
      "get" : {
        "tags" : [ "bod" ],
        "operationId" : "schema",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "businessObject",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/edis/sample-xml/{bod}" : {
      "get" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "sampleXml_1",
        "produces" : [ "application/xml" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/edis/sample-source-json/{bod}" : {
      "post" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "jsonSample_1",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/edis/sample-avro/{bod}" : {
      "get" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "sampleAvro_1",
        "produces" : [ "application/xml" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/edis/array-annotate/{bod}" : {
      "get" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "arrayAnnotate_1",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/edis/json-types/{bod}" : {
      "get" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "jsonTypes_1",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/edis/esql-gen/{bod}" : {
      "get" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "esqlGen_1",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/edis/construct/{bod}" : {
      "get" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "construct",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/edis/construct-annotate/{bod}" : {
      "get" : {
        "tags" : [ "Business Object Development Service" ],
        "operationId" : "constructAnnotate_1",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "bod",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/cmd/command/{group}/{command}" : {
      "get" : {
        "tags" : [ "Command Service" ],
        "operationId" : "command",
        "consumes" : [ "text/plain" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "group",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "command",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  }
}